refactor: extract types and functions for validations
