+++ src/interfaces/arrayOfCallbacks.type.ts
@@ -1,0 +1,1 @@
+export type TArrayOfCallbacks = (() => boolean)[];
