+++ src/interfaces/inputName.type.ts
@@ -1,0 +1,1 @@
+export type TInputName = "title" | "date";
