+++ src/shared/routerConfig.ts
@@ -1,4 +1,4 @@
-import { IRouterPath } from "@/types/routerPath.interface";
+import { IRouterPath } from "@/interfaces/routerPath.interface";
 
 export const ROUTER_CONFIG: IRouterPath[] = [
   {
