+++ src/shared/utils/validators/dateValidation.ts
@@ -1,0 +1,3 @@
+export function dateValidation(value: string): boolean {
+  return !isNaN(new Date(value).getTime());
+}
