+++ src/shared/utils/validators/validateAllFields.ts
@@ -1,0 +1,13 @@
+import { TArrayOfCallbacks } from "@/interfaces/arrayOfCallbacks.type";
+
+export function validateAllFields(arrayOfValidateFn: TArrayOfCallbacks): boolean {
+  let isAllValid = true;
+  arrayOfValidateFn.forEach((fn) => {
+    const isValid = fn();
+    if (!isValid) {
+      isAllValid = isValid;
+    }
+  });
+
+  return isAllValid;
+}
